feat: replacte matgen_value_t instead of f32 and f64

+++ include/matgen/core/types.h
@@ -54,10 +54,10 @@ typedef u64 matgen_index_t;
 /**
  * @brief Value type for matrix entries
  *
- * Currently using double precision (f64).
+ * Currently using matgen_value_t precision (f64).
  * Change to f32 for single precision, or make it templated.
  */
-typedef f64 matgen_value_t;
+typedef f32 matgen_value_t;
 
 /**
  * @brief Size type for counts (nnz, capacity, etc.)
